+++ lib/command/mail.rb
@@ -21,6 +21,7 @@ module Command
   ・<target>を省略した場合、新着があった小説を全て送信します。
   ・メールの送信設定は、#{Mailer::SETTING_FILE}ファイルを編集します。
     (初めてコマンドを使うときに自動で作成されます)
+  ・<target>にhotentryを指定した場合、最新のhotnetryを送信します
 
   Examples:
     narou mail 6         # 新着関係なくメール(送信済みフラグは立つ)
@@ -61,25 +62,33 @@ module Command
       end
       tagname_to_ids(argv)
       argv.each do |target|
-        ebook_path = Narou.get_ebook_file_path(target, device ? device.ebook_file_ext : ".epub")
+        if target == "hotentry"
+          ebook_path = Update.get_newest_hotentry_file_path(device)
+          display_target = "hotentry"
+        else
+          ebook_path = Narou.get_ebook_file_path(target, device ? device.ebook_file_ext : ".epub")
+          data = Downloader.get_data_by_target(target)
+          if send_all && !@options["force"]
+            new_arrivals_date = data["new_arrivals_date"] || Time.now
+            if data["last_mail_date"] && new_arrivals_date < data["last_mail_date"]
+              next   # すでに送信済みなので送信しない
+            end
+          end
+        end
         unless ebook_path
           error "#{target} は存在しません" unless send_all
           next
-        end
-        data = Downloader.get_data_by_target(target)
-        if send_all && !@options["force"]
-          new_arrivals_date = data["new_arrivals_date"] || Time.now
-          if data["last_mail_date"] && new_arrivals_date < data["last_mail_date"]
-            next   # すでに送信済みなので送信しない
-          end
         end
         unless File.exist?(ebook_path)
           error "まだファイル(#{File.basename(ebook_path)})が無いようです" unless send_all
           next
         end
-        id = data["id"]
-        title = data["title"]
-        puts "<bold><green>ID:#{id}　#{TermColorLight.escape(title)}</green></bold>".termcolor
+        if target != "hotentry"
+          id = data["id"]
+          title = data["title"]
+          display_target = "ID:#{id}　#{TermColorLight.escape(title)}"
+        end
+        puts "<bold><green>#{display_target}</green></bold>".termcolor
         print "メールを送信しています"
         exit_mail = false
         mail_result = nil
@@ -94,7 +103,7 @@ module Command
         puts
         if mail_result
           puts File.basename(ebook_path) + " をメールで送信しました"
-          database[id]["last_mail_date"] = Time.now
+          database[id]["last_mail_date"] = Time.now if target != "hotentry"
         else
           error "#{mailer.error_message}"
           exit Narou::EXIT_ERROR_CODE   # next しても次も失敗する可能性が高いのでここで終了
@@ -115,7 +124,9 @@ module Command
       puts "created #{install_path}"
       puts "メールの設定用ファイルを作成しました。設定ファイルを書き換えることで mail コマンドが有効になります。"
       puts "注意：次回以降のupdateで新着があった場合に送信可能フラグが立ちます"
-      Helper.open_directory(Narou.get_root_dir, "設定ファイルがあるフォルダを開きますか")
+      unless Narou.web?
+        Helper.open_directory(Narou.get_root_dir, "設定ファイルがあるフォルダを開きますか")
+      end
     end
 
     def alter_database_add_column_last_mail_date